# A draft PR to address some issues I've had (generated code)