--- a/examples/directives/.cargo/config.toml
+++ b/examples/directives/.cargo/config.toml
@@ -1,2 +1,0 @@
-[build]
-rustflags = ["--cfg=web_sys_unstable_apis"]