--- a/hydration_fix_tests/src/hydration_fix_validation.rs
+++ b/hydration_fix_tests/src/hydration_fix_validation.rs
@@ -1,0 +1,196 @@
+// Hydration Fix Validation Tests
+// Tests for the Leptos 0.8.x hydration tuple mismatch fix
+
+use leptos::prelude::*;
+
+#[cfg(test)]
+mod hydration_tuple_tests {
+    use super::*;
+
+    /// Test empty view compilation
+    /// Should compile without errors and not generate tuples
+    #[test]
+    fn test_empty_view() {
+        let _view = view! { };
+        // Empty view should compile successfully
+    }
+
+    /// Test single element view
+    /// Should generate single element, not tuple
+    #[test]
+    fn test_single_element_view() {
+        let _view = view! { <div>"Single element"</div> };
+        // Single element should not be wrapped in tuple
+    }
+
+    /// Test two element view  
+    /// Should generate 2-element tuple or be handled appropriately
+    #[test]
+    fn test_two_element_view() {
+        let _view = view! {
+            <div>"First"</div>
+            <span>"Second"</span>
+        };
+        // Two elements should be handled correctly
+    }
+
+    /// Test three element view
+    /// Should generate proper 3-element tuple - this is the expected case
+    #[test]
+    fn test_three_element_view() {
+        let _view = view! {
+            <div>"First"</div>
+            <span>"Second"</span>  
+            <p>"Third"</p>
+        };
+        // Three elements should generate proper 3-tuple
+    }
+
+    /// Test five element view - THE CRITICAL FAILING CASE
+    /// This is the specific case that fails with "expected 3 elements, found 5"
+    #[test]
+    fn test_five_element_view() {
+        let _view = view! {
+            <div>"First"</div>
+            <div>"Second"</div>
+            <div>"Third"</div>
+            <div>"Fourth"</div>
+            <div>"Fifth"</div>
+        };
+        // Five elements should be handled without tuple mismatch error
+        // This test will fail BEFORE the fix and pass AFTER
+    }
+
+    /// Test large view with many elements
+    /// Should handle >16 elements with existing chunking logic
+    #[test]
+    fn test_large_view() {
+        let _view = view! {
+            <div>"1"</div> <div>"2"</div> <div>"3"</div> <div>"4"</div> <div>"5"</div>
+            <div>"6"</div> <div>"7"</div> <div>"8"</div> <div>"9"</div> <div>"10"</div>
+            <div>"11"</div> <div>"12"</div> <div>"13"</div> <div>"14"</div> <div>"15"</div>
+            <div>"16"</div> <div>"17"</div> <div>"18"</div> <div>"19"</div> <div>"20"</div>
+        };
+        // Large view should use existing >16 element chunking
+    }
+
+    /// Test mixed static and dynamic content
+    /// Ensure tuple generation handles reactive content correctly
+    #[test]
+    fn test_mixed_content_view() {
+        let (signal, _set_signal) = signal(42);
+        
+        let _view = view! {
+            <div>"Static content"</div>
+            <span>{move || signal.get()}</span>
+            <p>"More static"</p>
+            <button on:click=move |_| {}>
+                "Interactive"
+            </button>
+            <input type="text" />
+        };
+        // Mixed content should be handled correctly
+    }
+
+    /// Test nested components
+    /// Ensure fix works with component hierarchies
+    #[test]
+    fn test_nested_components() {
+        #[component]
+        fn Inner() -> impl IntoView {
+            view! {
+                <div>"Inner component"</div>
+                <span>"Nested content"</span>
+            }
+        }
+
+        let _view = view! {
+            <div>"Outer"</div>
+            <Inner />
+            <p>"After component"</p>
+        };
+        // Nested components should work correctly
+    }
+}
+
+#[cfg(test)]
+mod feature_flag_tests {
+    use super::*;
+
+    /// Test CSR feature with hydration fix
+    #[test]
+    fn test_csr_five_elements() {
+        let _view = view! {
+            <div>"CSR First"</div>
+            <div>"CSR Second"</div>
+            <div>"CSR Third"</div>
+            <div>"CSR Fourth"</div>
+            <div>"CSR Fifth"</div>
+        };
+    }
+
+    /// Test SSR feature with hydration fix
+    #[test]
+    fn test_ssr_five_elements() {
+        let _view = view! {
+            <div>"SSR First"</div>
+            <div>"SSR Second"</div>
+            <div>"SSR Third"</div>
+            <div>"SSR Fourth"</div>
+            <div>"SSR Fifth"</div>
+        };
+    }
+
+    /// Test Hydrate feature with hydration fix (most critical)
+    #[test]
+    fn test_hydrate_five_elements() {
+        let _view = view! {
+            <div>"Hydrate First"</div>
+            <div>"Hydrate Second"</div>
+            <div>"Hydrate Third"</div>
+            <div>"Hydrate Fourth"</div>
+            <div>"Hydrate Fifth"</div>
+        };
+    }
+}
+
+#[cfg(test)]
+mod integration_tests {
+    use super::*;
+
+    /// Test the specific hydration module scenario that's failing
+    #[test]
+    fn test_hydration_mod_scenario() {
+        // This simulates the exact scenario from leptos/src/hydration/mod.rs:138
+        let root = "http://localhost:3000";
+        let pkg_path = "pkg";
+        let js_file_name = "app";
+        let wasm_file_name = "app";
+        let script = "import";
+        let islands_router = "";
+        let nonce = None::<String>;
+
+        let _view = view! {
+            <div>"Module preload"</div>
+            <div>"Preload"</div>
+            <div>"Script"</div>
+        };
+        // This should compile without tuple mismatch error
+    }
+
+    /// Test leptos-state compatibility scenario
+    #[test]
+    fn test_leptos_state_compatibility() {
+        // This test validates that the fix doesn't break leptos-state compatibility
+        // Note: This would require leptos-state as a dependency in a real scenario
+        
+        let _view = view! {
+            <div>"leptos-state compatible view"</div>
+            <span>"with multiple elements"</span>
+            <p>"that should work"</p>
+            <button>"after fix"</button>
+            <input type="text" />
+        };
+        // This should work with leptos-state hydration
+    }
+}