--- a/leptos_mode_cli/src/lib.rs
+++ b/leptos_mode_cli/src/lib.rs
@@ -1,0 +1,10 @@
+//! Leptos Mode CLI Library
+//!
+//! Library interface for the Leptos mode detection and migration CLI tool.
+
+pub mod commands;
+pub mod utils;
+
+// Re-export main functionality
+pub use commands::*;
+pub use utils::*;