… .github/workflows/test-matrix.yml
… CLAUDE.md
… Cargo.lock
… README.md
… design/leptos_init_strategy.md
… docs/API_REFERENCE.md
… docs/AUTOMATIC_MODE_DETECTION_GUIDE.md
… docs/AUTOMATIC_MODE_DETECTION_IMPLEMENTATION_SUMMARY.md
… docs/CONTRIBUTING.md
… docs/DEVELOPMENT_GUIDE.md
… docs/DEVELOPMENT_PERFORMANCE_IMPLEMENTATION_PLAN.md
… docs/FUTURE_IMPROVEMENTS_ROADMAP.md
… docs/HYDRATION_0.8_FIX_DESIGN.md
… docs/HYDRATION_FIX_IMPLEMENTATION_REPORT.md
… docs/HYDRATION_FIX_IMPLEMENTATION_SUMMARY.md
… docs/HYDRATION_FIX_PACKAGE_GUIDE.md
… docs/HYDRATION_TESTING_IMPLEMENTATION.md
… docs/LEPTOS_FAST_DEV_GUIDE.md
… docs/LEPTOS_INIT_DEVELOPER_GUIDE.md
… docs/LEPTOS_INIT_DOCUMENTATION_INDEX.md
… docs/LEPTOS_INIT_EXAMPLES.md
… docs/LEPTOS_INIT_GUIDE.md
… docs/LEPTOS_INIT_QUICK_REFERENCE.md
… docs/PACKAGE_DISTRIBUTION_INDEX.md
… docs/PHASE2_PERFORMANCE_REPORT.md
… docs/README.md
… docs/TESTING_IMPROVEMENTS.md
… docs/TESTING_INFRASTRUCTURE_SUMMARY.md
… docs/TESTING_STRATEGY.md
… docs/improvements/ISSUE_TEMPLATE.md