@gbj 2025-09-09
If you have experienced specific issues, please open issues.

If you would like to make PRs, please open PRs that
1) address a single issue, which is either linked or described in the PR in a way that can be reproduced,
2) contain code that you have *read* and can understand, even if an LLM *wrote* it
3) do not contain extraneous material like dozens of markdown files containing 

It is a waste of *your* time to open PRs like this because they are of such a massive scope that they are not useful.


I have not read all 50k lines of this diff or looked at all 177 files, but even a brief glance shows that this reproduces some of the major problems with #4275.

Some of the new material (like compile-time diagnostics for common issues!) sounds cool but the lints themselves are garbage:
```rust
        let help_dynamic = format!("For dynamic content, use: `{{move || {}.get()}}`", signal_name);
        let help_static = format!("For one-time reads, use: `{{{}.get()}}`", signal_name);
```
Neither of these is correct.